+++ juniper/src/types/name.rs
@@ -1,25 +1,30 @@
-use std::{
-    borrow::Borrow,
-    error::Error,
-    fmt::{Display, Formatter, Result as FmtResult},
-};
+use std::{borrow::Borrow, error::Error, fmt};
 
 use arcstr::ArcStr;
 
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct Name(ArcStr);
 
 impl Name {
-    pub fn new(input: ArcStr) -> Result<Self, NameParseError> {
-        if Self::is_valid(&input) {
-            Ok(Name(input))
+    /// Creates a new [`Name`] out of the provided `input` string, if it [`is_valid`].
+    ///
+    /// [`is_valid`]: Name::is_valid
+    pub fn new<S>(input: S) -> Result<Self, NameParseError>
+    where
+        S: AsRef<str> + Into<ArcStr>,
+    {
+        if Self::is_valid(input.as_ref()) {
+            Ok(Name(input.into()))
         } else {
-            Err(NameParseError(format!(
-                "Names must match /^[_a-zA-Z][_a-zA-Z0-9]*$/ but \"{input}\" does not",
+            Err(NameParseError(arcstr::format!(
+                "`Name` must match /^[_a-zA-Z][_a-zA-Z0-9]*$/ but \"{}\" does not",
+                input.as_ref(),
             )))
         }
     }
 
+    /// Validates the provided `input` string to represent a valid [`Name`[.
+    #[must_use]
     pub fn is_valid(input: &str) -> bool {
         for (i, c) in input.chars().enumerate() {
             let is_valid = c.is_ascii_alphabetic() || c == '_' || (i > 0 && c.is_ascii_digit());
@@ -31,21 +36,6 @@ impl Name {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct NameParseError(String);
-
-impl Display for NameParseError {
-    fn fmt(&self, f: &mut Formatter) -> FmtResult {
-        self.0.fmt(f)
-    }
-}
-
-impl Error for NameParseError {
-    fn description(&self) -> &str {
-        &self.0
-    }
-}
-
 impl Borrow<ArcStr> for Name {
     fn borrow(&self) -> &ArcStr {
         &self.0
@@ -54,6 +44,21 @@ impl Borrow<ArcStr> for Name {
 
 impl Borrow<str> for Name {
     fn borrow(&self) -> &str {
+        &self.0
+    }
+}
+
+#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
+pub struct NameParseError(ArcStr);
+
+impl fmt::Display for NameParseError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.0.fmt(f)
+    }
+}
+
+impl Error for NameParseError {
+    fn description(&self) -> &str {
         &self.0
     }
 }