+++ juniper_codegen/src/common/deprecation.rs
@@ -108,8 +108,8 @@ impl Directive {
 impl ToTokens for Directive {
     fn to_tokens(&self, into: &mut TokenStream) {
         let reason = self.reason.as_ref().map_or_else(
-            || quote! { None },
-            |text| quote! { Some(::juniper::literal!(#text)) },
+            || quote! { <::juniper::ArcStr>::None },
+            |text| quote! { Some(::juniper::arcstr::literal!(#text)) },
         );
         quote! {
             .deprecated(::core::option::Option::#reason)
@@ -134,7 +134,7 @@ mod parse_from_deprecated_attr_test {
                 .into_inner();
         assert_eq!(
             quote! { #desc }.to_string(),
-            quote! { .deprecated(::core::option::Option::Some(::juniper::literal!("foo"))) }
+            quote! { .deprecated(::core::option::Option::Some(::juniper::arcstr::literal!("foo"))) }
                 .to_string(),
         );
     }
@@ -147,7 +147,7 @@ mod parse_from_deprecated_attr_test {
             .into_inner();
         assert_eq!(
             quote! { #desc }.to_string(),
-            quote! { .deprecated(::core::option::Option::None) }.to_string(),
+            quote! { .deprecated(::core::option::Option::<::juniper::ArcStr>::None) }.to_string(),
         );
     }