+++ juniper_codegen/src/common/field/mod.rs
@@ -299,7 +299,8 @@ impl Definition {
             .flat_map(|args| args.iter().filter_map(MethodArgument::method_meta_tokens));
 
         quote! {
-            registry.field_convert::<#ty, _, Self::Context>(::juniper::literal!(#name), info)
+            registry
+                .field_convert::<#ty, _, Self::Context>(::juniper::arcstr::literal!(#name), info)
                 #( #args )*
                 #description
                 #deprecated