+++ juniper_codegen/src/graphql_union/mod.rs
@@ -473,17 +473,14 @@ impl Definition {
             #[automatically_derived]
             impl #impl_generics ::juniper::GraphQLType<#scalar> for #ty_full #where_clause
             {
-                fn name(
-                    _ : &Self::TypeInfo,
-                ) -> ::core::option::Option<::juniper::ArcStr> {
-                    ::core::option::Option::Some(::juniper::literal!(#name))
+                fn name(_ : &Self::TypeInfo) -> ::core::option::Option<::juniper::ArcStr> {
+                    ::core::option::Option::Some(::juniper::arcstr::literal!(#name))
                 }
 
                 fn meta(
                     info: &Self::TypeInfo,
-                    registry: &mut ::juniper::Registry<#scalar>
-                ) -> ::juniper::meta::MetaType<#scalar>
-                {
+                    registry: &mut ::juniper::Registry<#scalar>,
+                ) -> ::juniper::meta::MetaType<#scalar> {
                     let types = [
                         #( registry.get_type::<#variant_tys>(info), )*
                     ];
@@ -526,9 +523,9 @@ impl Definition {
                 type Context = #context;
                 type TypeInfo = ();
 
-                fn type_name<'__i>(
+                fn type_name(
                     &self,
-                    info: &'__i Self::TypeInfo,
+                    info: &Self::TypeInfo,
                 ) -> ::core::option::Option<::juniper::ArcStr> {
                     <Self as ::juniper::GraphQLType<#scalar>>::name(info)
                 }